Rename the contrib.migrations module to contrib.migration.
